Updated the hasMany relation to return actionable objects.
The addition/removal/clearing of objects now simply stores information
about the desired changes to be made to the models. This is then acted
upon during save.

The reason for this change is that it makes the hasMany and belongsTo
relationships far more congruent with respect to when data is actually
persisted.