Pluralization not required.
At this point, the source of a through relation will never require
pluralization. The name of the relationship will be plural since
has-many is the only thing that accepts a `through`. The final source
may eventually be singular (a belongs-to), but we can't go through a
belongs-to and back to a has-many. This may eventually be needed if
adding has-one support and giving that a `through` option.