An attempt to simplify pre-fetch through.
This attempt was based on the idea of performing fewer queries during
a through pre-fetch, but ended up not being a good idea for performance
reasons. Duplicate data would have to be returned during `LEFT` or
`RIGHT` joins that would need to be performed in order to achieve this
result.

This commit is being kept for the sake of being able to look back at it
if in the future we decide to try to support a pre-fetch of this style,
but is not really involved in the process of making the current
many-to-many feature set work.