Models give a better warning about being set up.
This isn't quite ideal because it doesn't address failing to set up the
model properly in any case but insert. It is possible, however, that
someone may want to create a model that only has a primary key. Making
it an error to create a model with just a primary key doesn't seem like
it should be a hard error (though it's certainly silly).

This fixes #11.