Always lookup relation object on model instance.
All methods and properties that are added by a relation are now
required to look up the relation object when they're invoked.
Previously, the relation object used to define the method was assumed
to be constant, but the lookup allows the relation object to be
overridden & redefine aspects of the relation object.

This is required for supporting relations in azul-transaction.