Overhaul of how through relations are formed.
Through relations now represent themselves and do not need an implicit
(or explicit) relation to which they relate. This means that we don't
need to decide any more if we want to add a hasMany relation when
creating a through relation.

The advantage of this is that fewer implicit relations are added which
allowed mis-configurations to easily go overlooked.